+++ actix-test/CHANGES.md
@@ -3,8 +3,9 @@
 ## Unreleased
 
 - Add `TestServerConfig::rustls_0_23()` method for Rustls v0.23 support behind new `rustls-0_23` crate feature.
-- Minimum supported Rust version (MSRV) is now 1.72.
+- Add `TestServerConfig::disable_redirects()` method.
 - Various types from `awc`, such as `ClientRequest` and `ClientResponse`, are now re-exported.
+- Minimum supported Rust version (MSRV) is now 1.72.
 
 ## 0.1.3