+++ actix-web-codegen/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- Prevent inclusion of default `actix-router` features.
 - Minimum supported Rust version (MSRV) is now 1.72.
 - Add a scope macro that takes a path